test: css, js 정적 파일을 읽어오는 테스트 추가
